chore: create separate meiliSync logger, clean up logs to avoid being unnecessarily verbose
